pyproject.toml
[tool.poetry]
name = "todo-backend"
version = "0.0.1"
description = "A Django-based REST API for a todo application"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.16.0"
django-allauth = "^65.8.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
